// src/domains/domain-auth/authentications/enums/authentication-messages.enum.ts
export enum AuthenticationErrors {
  INCORRECT_VERIFICATION_CODE = 'authentication.INCORRECT_VERIFICATION_CODE',
  INCORRECT_2FA_TYPE = 'authentication.INCORRECT_2FA_TYPE',
  TWO_FA_METHOD_ALREADY_SET = 'authentication.TWO_FA_METHOD_ALREADY_SET',
  TWO_FA_METHOD_NOT_SET = 'authentication.TWO_FA_METHOD_NOT_SET',
  VERIFICATION_CODE_ALREADY_SENT = 'authentication.VERIFICATION_CODE_ALREADY_SENT',
  FORGET_PASSWORD_CODE_ALREADY_SENT = 'authentication.FORGET_PASSWORD_CODE_ALREADY_SENT',
  INVALID_TOKEN = 'authentication.INVALID_TOKEN',
  INVALID_CAPTCHA = 'authentication.INVALID_CAPTCHA',
  NOT_FOUND_SESSION = 'authentication.NOT_FOUND_SESSION',
}